SO! Fixed issue of reupdating data when "back" is pressed
Apparently react navigation is actually dumb! you have to use some hook to actually update the table data in aws when you click "back" from interests and try to edit goals then, otherwise it doesn't update!! thank god i figured out this was an issue cause i thought it was a 1 time thing. See Tim's comment on the top vote answer: https://stackoverflow.com/questions/60182942/useeffect-not-called-in-react-native-when-back-to-screen