OMG i can't believe AWS IS LIKE THIS!!!
oK SO! Look at the changes in OnboardingGoals.js. Basically I would delete each insight for a certain goal in a function, but it only had one try catch. I told them to only delete insights with status "new". If you try to delete an insight with "current" it fails and causes an AWS error. Since I was trying to delete multiple insights in the function, if the AWS error happened, the other insights wouldn't be deleted (if i was trying to delete the 8 Insights of "Reduce Anxiety", it would stop deleting as soon as it hit one insight that was current, so the rest of the Reduce Anxiety insights wouldn't be deleted). So I moved ther try-catch to inside the loop - where you go thru and delete all the Reduce Anxiety insights, so that even if it fails for one, it stays in the loop. Again, this failure isn't a real failure, because it's just a current insight, but we should delete the other insights of the goal we just deleted. I should learn about delete resolvers but it's hard lol (would also help with the limit thing if we're off the 1st page): https://docs.aws.amazon.com/appsync/latest/devguide/tutorial-dynamodb-resolvers.html#setting-up-the-deletepost-resolver-ddb-deletepost